Math expression
I propose (what I suspect) to be a more mathematical way to express the idea that we are trying to do. I believe this shows what we intend to happen better than hardcoding the number halves. Still not quite modular, but I believe it helps get people closer to thinking in a modular way with this version.